Update Docker setup and add user login functionality
The Dockerfile for the frontend app was removed and the main Dockerfile was updated to use the newer python image and enable unbuffered output. The `requirements.txt` file was also updated with new dependencies. New user login and registration functionality, along with corresponding HTML pages were added to `app.py`. In addition, the 'weather-widget.html' file was moved to the frontend directory.